refactor(context-split): retire deleted-machinery prose + clippy nits (5b closeout)
Cleanliness sweep closing stage 5b. No behaviour change.

Stale doc references to machinery deleted earlier in 5b:
- `HostMut` had zero live code references but survived in doc comments and
  broken `[HostMut]` intra-doc links. Re-point the `[HostMut::verb]` mirrors in
  the pipeline `FunctionBody` surface at the concrete `Function::verb` they now
  delegate to (`instruction_mut` -> `Function::insn_mut`); reword the module-doc
  narration in module_view.rs / builder.rs (which still described the gone
  transitional `impl<T: HostMut> BuilderBacking` blanket) to the two-direct-impls
  reality; reword `HostMut` mentions in block.rs / base_ref.rs / cse.rs /
  memory.rs / dead_insns.rs / dead_block_args.rs / licm.rs.
- `checkout`/`check-in` driver-protocol prose (the driver borrows bodies in place
  via `split()` since c.3): "check-in" -> "the barrier", "checkout" -> "the
  split"/"before the run", "checkout-safe" -> "self-stored". The two "check-out
  shim" mentions naming the live `with_checked_out_body` function are kept.

Dead code + `cargo clippy --all-targets` mechanical fixes in 5b-touched code:
unused `context::Context` / `Value` test imports, redundant `i64`->`i64` casts,
a needless `&mut host` ref + consequent `unused_mut`, a `?`-operator rewrite,
`.to_vec()`, a needless deref, items-after-test-module, and a `drop()`-on-non-Drop
no-op.

+++ analysis/src/gvn/mem_forward.rs
@@ -775,11 +775,12 @@ mod tests {
         }
     }
 
-    /// Check `fid` out and run `f` against its `(&mut FunctionBody, ContextView)`
-    /// — the only surface [`MemForward::record_store_c`]/[`MemForward::try_load_c`]
-    /// speak now that the `&mut Context` module twins are gone. The `MemForward`
-    /// state under test is owned by the caller (captured by `f`), so it outlives
-    /// the checkout and can be inspected afterwards.
+    /// Borrow `fid`'s body in place and run `f` against its
+    /// `(&mut FunctionBody, ContextView)` — the only surface
+    /// [`MemForward::record_store_c`]/[`MemForward::try_load_c`] speak now that the
+    /// `&mut Context` module twins are gone. The `MemForward` state under test is
+    /// owned by the caller (captured by `f`), so it outlives the borrow and can be
+    /// inspected afterwards.
     fn with_body<R>(
         tc: &mut TestContext,
         fid: qcode::value::FunctionId,
@@ -912,9 +913,7 @@ mod tests {
             }
             !regs.iter().any(|&r| {
                 let vn = Varnode::from_id(&tc.ctx, r);
-                vn.space().id == sp
-                    && (vn.address() as i64) <= off
-                    && off < vn.address() as i64 + vn.size() as i64
+                vn.space().id == sp && vn.address() <= off && off < vn.address() + vn.size() as i64
             })
         });
